requirements.txt
faiss-cpu
sentence-transformers
spacy
requests
tqdm
numpy
transformers
torch
